small fixes of select count(*)
ndb/src/ndbapi/NdbDictionaryImpl.cpp:
  Fix init
ndb/tools/select_count.cpp:
  row count is 64 bit

+++ ndb/tools/select_count.cpp
@@ -144,7 +144,7 @@ select_count(Ndb* pNdb, const NdbDictionary::Table* pTab,
       return NDBT_FAILED;
     }
   
-    Uint32 tmp;
+    Uint64 tmp;
     pOp->getValue(NdbDictionary::Column::ROW_COUNT, (char*)&tmp);
     
     check = pTrans->execute(NoCommit);
